+++ src/constants.js
@@ -104,6 +104,7 @@ export default {
   PRICE_TAG_STATUS_UNREADABLE: 2,
   PRICE_TAG_STATUS_TRUNCATED: 3,
   PRICE_TAG_STATUS_NOT_A_PRICE: 4,
+  PRODUCT_IMAGE_DEFAULT_URL: 'https://world.openfoodfacts.org/images/icons/dist/packaging.svg',
   PRODUCT_CATEGORY_LABEL_ORGANIC: PRODUCT_CATEGORY_LABEL_ORGANIC,
   PRODUCT_QUANTITY_UNIT_G: 'g',
   PRODUCT_QUANTITY_UNIT_ML: 'ml',
@@ -213,6 +214,11 @@ export default {
     // { key: 'recent', value: 'Recent', valueSmallScreen: '', icon: 'mdi-history' },  // Recent
     { key: 'scan', value: 'BarcodeScan', valueSmallScreen: 'BarcodeScanShort', icon: 'mdi-barcode-scan' },
     { key: 'type', value: 'BarcodeType', valueSmallScreen: 'BarcodeTypeShort', icon: 'mdi-numeric' },
+  ],
+  BARCODE_SCANNER_DISPLAY_LIST: [
+    { key: 'auto', value: 'Auto', valueSmallScreen: 'Auto', icon: 'mdi-barcode-scan' },
+    { key: 'off-barcode-scanner', value: 'Off Barcode Scanner', valueSmallScreen: 'Off Scanner', icon: 'mdi-barcode-scan' },
+    { key: 'html5-qrcode', value: 'Html5-Qrcode', valueSmallScreen: 'Html5-Qrcode', icon: 'mdi-barcode-scan' },
   ],
   PRICE_FORM_DISPLAY_LIST: [
     { key: 'display', value: 'Display', icon: 'mdi-eye-outline' },