+++ src/views/ProductDetail.vue
@@ -45,7 +45,7 @@
       </v-row>
     </v-window-item>
     <v-window-item value="table">
-      <PriceTable class="mt-3" :priceList="priceList" />
+      <PriceTable class="mt-3 mb-3" :priceList="priceList" source="product" />
     </v-window-item>
     <v-window-item value="map">
       <v-row class="mt-0 mb-1">
@@ -115,7 +115,7 @@ export default {
   computed: {
     ...mapStores(useAppStore),
     productIsCategory() {
-      return this.productId.startsWith('en')
+      return this.productId.includes(':')
     },
     productNotFound() {
       return !this.productIsCategory && this.product && !this.product.source